pyproject.toml
[project]
name = "research-bot"
version = "2.0.0"
description = "Review paper database builder — search, screen, curate, export academic papers"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    # HTTP client
    "httpx>=0.28.0",
    # LLM
    "google-genai>=1.64.0",
    # Search sources
    "arxiv>=2.1.0",
    "scholarly>=1.7.0",
    # Data processing
    "pandas>=2.0.0",
    "rapidfuzz>=3.0.0",
    # Export
    "pyzotero>=1.5.0",
    # Config & utilities
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
